Combine CentreMarker with MarkerVisualisation
